backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python<=4.6.0.66
Pillow==10.3.0
pypdfium2==4.30.0
google-genai>=1.0.0
rapidfuzz==3.9.0
requests==2.31.0
scispacy>=0.5.4
medspacy==1.3.1
spacy>=3.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
groq>=1.2.0
numpy==1.26.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
loguru==0.7.2
tqdm==4.66.2
sqlalchemy==2.0.30
bcrypt==4.1.2
